+++ modules/shared/core/src/sap/commerce/toolset/settings/state/ImpExSettingsState.kt
@@ -28,8 +28,8 @@ data class ImpExSettingsState(
     @JvmField @OptionTag val completion: ImpExCompletionSettingsState = ImpExCompletionSettingsState(),
     @JvmField @OptionTag val documentation: ImpExDocumentationSettingsState = ImpExDocumentationSettingsState(),
     // Type name to > set of attribute names
-    @JvmField @OptionTag val quoteStringMatching: Boolean = true,
-    @JvmField @OptionTag val quoteStringMatchingRegex: Regex = Regex("^[a-zA-Z0-9_$/\\-]+$"),
+    @JvmField @OptionTag val quoteStringWhitelist: Boolean = true,
+    @JvmField @OptionTag val quoteStringWhitelistPattern: Regex = Regex("^[a-zA-Z0-9_$/\\-]+$"),
     @JvmField @OptionTag val quoteStringExclusions: Map<String, Set<String>> = mapOf(),
 ) {
     fun mutable() = Mutable(
@@ -37,8 +37,8 @@ data class ImpExSettingsState(
         editMode = editMode.mutable(),
         completion = completion.mutable(),
         documentation = documentation.mutable(),
-        quoteStringMatching = quoteStringMatching,
-        quoteStringMatchingPattern = quoteStringMatchingRegex.pattern,
+        quoteStringWhitelist = this@ImpExSettingsState.quoteStringWhitelist,
+        quoteStringWhitelistPattern = this@ImpExSettingsState.quoteStringWhitelistPattern.pattern,
         quoteStringExclusions = quoteStringExclusions
             .flatMap { (type, attributes) ->
                 attributes.map { ImpExQuoteStringExclusion(type, it) }
@@ -51,8 +51,8 @@ data class ImpExSettingsState(
         var editMode: ImpExEditModeSettingsState.Mutable,
         var completion: ImpExCompletionSettingsState.Mutable,
         var documentation: ImpExDocumentationSettingsState.Mutable,
-        var quoteStringMatching: Boolean,
-        var quoteStringMatchingPattern: String,
+        var quoteStringWhitelist: Boolean,
+        var quoteStringWhitelistPattern: String,
         var quoteStringExclusions: MutableList<ImpExQuoteStringExclusion>,
     ) {
         fun immutable() = ImpExSettingsState(
@@ -60,8 +60,8 @@ data class ImpExSettingsState(
             editMode = editMode.immutable(),
             completion = completion.immutable(),
             documentation = documentation.immutable(),
-            quoteStringMatching = quoteStringMatching,
-            quoteStringMatchingRegex = Regex(quoteStringMatchingPattern),
+            quoteStringWhitelist = quoteStringWhitelist,
+            quoteStringWhitelistPattern = Regex(quoteStringWhitelistPattern),
             quoteStringExclusions = quoteStringExclusions
                 .groupBy(
                     keySelector = { it.typeName },